requirements.txt
pytest-django
requests
pillow
sorl-thumbnail
django
djangorestframework
django-cors-headers